+++ tests/functional/test_admin_org_endpoints.py
@@ -83,11 +83,11 @@ async def test_admin_can_create_org_with_new_email(app, client):
         ):
             with patch("app.auth.deps.get_or_create_user", return_value=admin):
                 with patch(
-                    "app.services.invite_service.user_service.find_user_by_email",
+                    "app.invites.service.user_service.find_user_by_email",
                     return_value=None,
                 ):
                     with patch(
-                        "app.services.invite_service.invite_cognito_user"
+                        "app.invites.service.invite_cognito_user"
                     ) as mock_cognito:
                         with patch(
                             "app.users.service.create_placeholder_user",
@@ -134,7 +134,7 @@ async def test_admin_can_create_org_with_existing_user(app, client):
                         return_value=None,
                     ):
                         with patch(
-                            "app.services.invite_service.invite_cognito_user"
+                            "app.invites.service.invite_cognito_user"
                         ) as mock_cognito:
                             with patch("app.orgs.service.create_membership"):
                                 response = await client.post(
